+++ internal/provider/model_sdwan_custom_control_topology_policy_definition.go
@@ -31,6 +31,7 @@ import (
 type CustomControlTopologyPolicyDefinition struct {
 	Id            types.String                                     `tfsdk:"id"`
 	Version       types.Int64                                      `tfsdk:"version"`
+	Type          types.String                                     `tfsdk:"type"`
 	Name          types.String                                     `tfsdk:"name"`
 	Description   types.String                                     `tfsdk:"description"`
 	DefaultAction types.String                                     `tfsdk:"default_action"`