+++ internal/provider/resource_sdwan_hub_and_spoke_topology_policy_definition.go
@@ -71,6 +71,10 @@ func (r *HubAndSpokeTopologyPolicyDefinitionResource) Schema(ctx context.Context
 				MarkdownDescription: "The version of the object",
 				Computed:            true,
 			},
+			"type": schema.StringAttribute{
+				MarkdownDescription: "Type",
+				Computed:            true,
+			},
 			"name": schema.StringAttribute{
 				MarkdownDescription: helpers.NewAttributeDescription("The name of the policy definition").String,
 				Required:            true,
@@ -192,6 +196,7 @@ func (r *HubAndSpokeTopologyPolicyDefinitionResource) Create(ctx context.Context
 	}
 	plan.Id = types.StringValue(res.Get("definitionId").String())
 	plan.Version = types.Int64Value(0)
+	plan.Type = types.StringValue("hubAndSpoke")
 
 	tflog.Debug(ctx, fmt.Sprintf("%s: Create finished successfully", plan.Name.ValueString()))